basic tabs for create join table, misc refactor
