+++ src/components/table/tableSelection/CreateTable.tsx
@@ -1,0 +1,6 @@
+export default function CreateTable() {
+  return (
+    <>
+    </>
+  )
+}
