+++ src/components/table/tableSelection/JoinTable.tsx
@@ -1,0 +1,6 @@
+export default function JoinTable() {
+  return (
+    <>
+    </>
+  )
+}
