+++ src/layouts/tableSelected/RestaurantListLayout.tsx
@@ -1,4 +1,4 @@
-import RestaurantPreview from "@/components/table/RestaurantPreview";
+import RestaurantPreview from "@/components/table/tableSelected/RestaurantPreview";
 import { Title } from "@mantine/core";
 
 export default function RestaurantListLayout() {
